quieting clang-tidy, use unsigned
places where we shouldn't be negative anyway

+++ modules/libcom/src/as/asLibRoutines.c
@@ -366,7 +366,7 @@ long epicsStdCall asAddClient(ASCLIENTPVT *pasClientPvt,ASMEMBERPVT asMemberPvt,
 {
     ASGMEMBER   *pasgmember = asMemberPvt;
     ASGCLIENT   *pasgclient;
-    int         len, i;
+    size_t      len, i;
 
     long        status;
     if(!asActive) return(S_asLib_asNotActive);
@@ -394,7 +394,7 @@ long epicsStdCall asChangeClient(
 {
     ASGCLIENT   *pasgclient = asClientPvt;
     long        status;
-    int         len, i;
+    size_t      len, i;
 
     if(!asActive) return(S_asLib_asNotActive);
     if(!pasgclient) return(S_asLib_badClient);
